don't try to link every library in MPI_PREFIX
Don't try to link every library in the MPI_PREFIX directory.  It's unnecessary
and it may cause the build to fail.  E.g. a standard Ubuntu install of MPICH
is to /usr/lib - linking every library in this directory can cause problems
for example with libpreloadable_intl.so on my machine.
Linking only the mpi libraries seems to work OK.